refactor: simplify study model controls
